refactor: move/deprecate components
