+++ apps/hash-frontend/src/shared/edit-icon-button/emoji-picker/emoji-picker.tsx
