+++ apps/hash-frontend/src/shared/use-update-page-icon.ts
@@ -2,12 +2,12 @@ import { useMutation } from "@apollo/client";
 import { EntityId } from "@local/hash-subgraph";
 import { useCallback } from "react";
 
+import { useGetPageRefetchQueries } from "../components/hooks/shared/get-page-refetch-queries";
 import {
   UpdatePageMutation,
   UpdatePageMutationVariables,
-} from "../../graphql/api-types.gen";
-import { updatePage } from "../../graphql/queries/page.queries";
-import { useGetPageRefetchQueries } from "./shared/get-page-refetch-queries";
+} from "../graphql/api-types.gen";
+import { updatePage } from "../graphql/queries/page.queries";
 
 export const useUpdatePageIcon = () => {
   const [updatePageFn, { loading: updatePageIconLoading }] = useMutation<
